+++ internal/generator/constructor/doc.go
@@ -1,2 +1,0 @@
-// Package constructor generates a "constructor" function for each message type.
-package constructor