+++ internal/generator/mutator/fielddiscrete.go
@@ -31,33 +31,12 @@ func generateForDiscreteField(code *jen.File, f *scope.Field) {
 				Id("v").
 				Add(f.GoType()),
 		).
-		Params(
-			jen.
-				Op("*").
-				Id(f.Message.GoTypeName),
-		).
+		Params().
 		Block(
-			jen.If(
-				jen.
-					Id("x").
-					Op("==").
-					Nil(),
-			).
-				Block(
-					jen.
-						Id("x").
-						Op("=").
-						Op("&").
-						Id(f.Message.GoTypeName).
-						Values(),
-				),
 			jen.
 				Id("x").
 				Dot(f.GoFieldName).
 				Op("=").
 				Id("v"),
-			jen.
-				Return().
-				Id("x"),
 		)
 }