+++ CHANGELOG.md
@@ -10,7 +10,7 @@ The format is based on [Keep a Changelog], and this project adheres to
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
 
-## [Unreleased]
+## [0.2.0] - 2023-10-19
 
 ### Added
 
@@ -76,6 +76,7 @@ The format is based on [Keep a Changelog], and this project adheres to
 [0.1.3]: https://github.com/dogmatiq/primo/releases/tag/v0.1.3
 [0.1.4]: https://github.com/dogmatiq/primo/releases/tag/v0.1.4
 [0.1.5]: https://github.com/dogmatiq/primo/releases/tag/v0.1.5
+[0.2.0]: https://github.com/dogmatiq/primo/releases/tag/v0.2.0
 
 <!-- version template
 ## [0.0.1] - YYYY-MM-DD